# Haulory - feat(jobs): add job creation, assignment and job list details
- CreateJobCommand/Handler with SortOrder + invoice generation
- Job entity billing + assignment fields
- Driver repository: GetById + GetByUserId + Save
- Jobs list UI: show driver/truck/pickup/drop + rate/total
- NewJob flow: driver/vehicle pickers + billing calculation
- EF Core model updates for job/driver/vehicle relationships and indices

Haulory is a .NET MAUI logistics app for owner-drivers and transport operators.
It supports creating Jobs (pickup → delivery), assigning Drivers and Vehicles, calculating billing totals, and capturing delivery signatures.

## Tech
- .NET MAUI (Mobile UI)
- Clean-ish layering:
  - `Haulory.Domain` (Entities + Enums)
  - `Haulory.Application` (Commands/Handlers + Interfaces)
  - `Haulory.Infrastructure` (EF Core + Repositories)
  - `Haulory.Mobile` (Pages + ViewModels)

## Current Features (this commit)
### Jobs
- Create a Job with:
  - Pickup + Delivery details
  - Reference + load description
  - Billing rate type/value/quantity with computed total
  - Optional assignment: Driver + VehicleAsset
- Jobs list page:
  - Displays driver + truck + pickup + drop + billing summary
  - Supports re-ordering jobs via SortOrder
  - Navigates to signature capture page (DeliverySignaturePage) by JobId

### Drivers
- Drivers belong to an OwnerUserId (main account).
- A Driver may optionally link to a main user account via `UserId` (main profile).
- Driver repository includes:
  - Get by OwnerUserId
  - Get by UserId
  - Get by Id
  - Save (create/update)

### Vehicles
- Vehicle assets support:
  - Power units + trailers using UnitNumber slots (1 = power unit, 2/3 = trailers)
  - Compliance: Rego expiry, COF/WOF expiry
  - RUC support (diesel/electric + heavy trailers)

### Delivery receipts
- `DeliveryReceipt` stores a completed delivery snapshot including signature JSON.

## Quick Start
1. Ensure EF Core migrations are applied (SQLite).
2. Run the MAUI project.
3. Create Drivers and Vehicles.
4. Create Jobs and assign driver/vehicle.
5. Tap **Sign Delivery** to capture proof of delivery.

# Jobs

A Job represents a single delivery run: Pickup → Delivery with billing and optional assignment.

## Domain: `Job`
File: `Haulory.Domain.Entities.Job`

### Core Fields
- `OwnerUserId` (required): main user who owns the job
- `AssignedToUserId` (optional): reserved for sub-user assignment later
- `PickupCompany`, `PickupAddress`
- `DeliveryCompany`, `DeliveryAddress`
- `ReferenceNumber`
- `LoadDescription`

### Assignment
- `DriverId` (optional)
- `VehicleAssetId` (optional)

### Billing
- `InvoiceNumber`
- `RateType` (enum)
- `RateValue` (decimal)
- `Quantity` (decimal)
- `QuantityUnit` (derived from RateType)
- `Total` (computed):
  - FixedFee => `RateValue`
  - otherwise => `RateValue * Quantity`

### Delivery completion
- `ReceiverName`
- `DeliveredAtUtc`
- `DeliverySignatureJson`
- `IsDelivered` is true when both DeliveredAtUtc and signature JSON exist.

### Ordering
- `SortOrder` controls the display order in the Jobs list.
- Re-ordering updates SortOrder and persists via repository.

---

## Creating a Job

### Command
File: `Haulory.Application.Features.Jobs.CreateJobCommand`

```csharp
public record CreateJobCommand(
    Guid OwnerUserId,
    string PickupCompany,
    string PickupAddress,
    string DeliveryCompany,
    string DeliveryAddress,
    string ReferenceNumber,
    string LoadDescription,
    RateType RateType,
    decimal RateValue,
    decimal Quantity,
    Guid? DriverId,
    Guid? VehicleAssetId
);
# Vehicles (VehicleAsset)

VehicleAsset supports powered vehicles + trailers using slot-based rules.

## Domain: `VehicleAsset`
File: `Haulory.Domain.Entities.VehicleAsset`

### Slot Rules
- Unit 1 = Power Unit (Vehicle)
- Unit 2 = Trailer 1
- Unit 3 = Trailer 2 (B-Train)

### Ownership
- `OwnerUserId` is required and is used as the foreign key to UserAccount.

### Compliance
- `Rego` + `RegoExpiry`
- `CertificateType` (WOF/COF) + `CertificateExpiry`
- Convenience fields:
  - Expiry display strings
  - Due soon / expired flags

### RUC (Road User Charges)
Stored per asset (per slot).
Applies to:
- Power unit if Diesel/Electric
- Heavy trailers (Class 3 & Class 5)

Includes:
- Purchase metadata
- Licence range start/end
- Remaining km based on current odometer

## UI: Vehicle Wizard
ViewModel: `Haulory.Mobile.ViewModels.NewVehicleViewModel`

This wizard:
- Enforces slot rules (which units must be filled)
- Displays and validates per-unit rego, expiry, make/model/year, cert expiry
- Handles odometer requirements by vehicle type
- Adds RUC step only when it applies

Saving:
- Powered vehicles: creates Unit 1 asset
- Trailer-only: creates Unit 2 (+ Unit 3 if B-Train)
# Database Model (EF Core)

DbContext:
File: `Haulory.Infrastructure.Persistence.HauloryDbContext`

DbSets:
- `UserAccounts`
- `Drivers`
- `Jobs`
- `VehicleAssets`
- `DeliveryReceipts`

## Relationships

### Driver
- Required: Driver.OwnerUserId -> UserAccount (Cascade delete)
- Optional: Driver.UserId -> UserAccount (Restrict delete)
- Owned type: EmergencyContact

### VehicleAsset
- Required: VehicleAsset.OwnerUserId -> UserAccount (Cascade delete)
- Indexes:
  - OwnerUserId
  - VehicleSetId
  - Rego
  - Unique: (VehicleSetId, UnitNumber)

### Job
- Required: Job.OwnerUserId -> UserAccount (Cascade delete)
- Optional: Job.DriverId -> Driver (SetNull)
- Optional: Job.VehicleAssetId -> VehicleAsset (SetNull)
- Decimal columns:
  - Quantity, RateValue => decimal(18,2)
- Indexes:
  - OwnerUserId
  - DriverId
  - VehicleAssetId

### DeliveryReceipt
- 1 receipt per Job:
  - Unique index on JobId
- FK: DeliveryReceipt.JobId -> Job (Cascade delete)
- Decimal columns: RateValue, Total => decimal(18,2)
- Required: SignatureJson, ReceiverName
# UI Pages & Flow (MAUI)

## JobsCollectionPage
File: `Haulory.Mobile.Views.JobsCollectionPage`

- Shows list of active jobs.
- Uses `JobsCollectionViewModel`.
- Calls `LoadAsync()` in `OnAppearing`.

### Actions
- Add Job -> navigates to `NewJobPage`
- Reorder: ▲ / ▼ updates SortOrder and persists
- Sign Delivery -> navigates to `DeliverySignaturePage?jobId=...`

## NewJobPage
File: `Haulory.Mobile.Views.NewJobPage`

ViewModel: `Haulory.Mobile.ViewModels.NewJobViewModel`

Fields captured:
- Pickup company/address
- Delivery company/address
- Reference number
- Load description
- Assignment:
  - Driver picker (DisplayName)
  - Vehicle picker (Rego)
- Billing:
  - RateType, RateValue, Quantity
  - Total computed live

OnAppearing:
- Calls `LoadAsync()` to load Drivers and Vehicles for current OwnerUserId (session).

Saving:
- `SaveJobCommand` calls CreateJobHandler and returns to JobsCollectionPage.

## DriverCollectionPage (Drivers)
ViewModel: `Haulory.Mobile.ViewModels.DriverCollectionViewModel`

- Restores session if needed
- Loads drivers owned by main user
- Ensures the main profile driver exists:
  - If missing, creates it from UserAccount via CreateDriverFromUserHandler
- Controls UI gating:
  - Continue button until main driver emergency contact is complete
  - Add driver appears once main profile is complete

+++ Haulory.Domain/Entities/DeliveryReceipt.cs
@@ -19,7 +19,7 @@ public class DeliveryReceipt
 
     public RateType RateType { get; private set; }
     public decimal RateValue { get; private set; }
-    public int Quantity { get; private set; }
+    public decimal Quantity { get; private set; }
     public decimal Total { get; private set; }
 
     public string ReceiverName { get; private set; } = string.Empty;
@@ -27,7 +27,7 @@ public class DeliveryReceipt
 
     public string SignatureJson { get; private set; } = string.Empty;
 
-    // ✅ Required by EF
+    // Required by EF
     public DeliveryReceipt() { }
 
     public DeliveryReceipt(
@@ -41,7 +41,7 @@ public class DeliveryReceipt
         string loadDescription,
         RateType rateType,
         decimal rateValue,
-        int quantity,
+        decimal quantity,
         decimal total,
         string receiverName,
         DateTime deliveredAtUtc,
